WIP: patterns can reference tokens
The code now works, although it is not clear the feature
will be useful. The challenge is that a pattern like
  "@QUANTITY of them"
often matches every @QUANTITY token, e.g.
  "one red car"

It seems that you only want to allow token references in patterns where the matcher is configured to only allow
exact matches.